+++ NfcTag.h
@@ -9,22 +9,22 @@ class NfcTag
 {
     public:
         NfcTag();
-        NfcTag(uint8_t* uid, uint8_t uidLength);
-        NfcTag(uint8_t* uid, uint8_t uidLength, String tagType);
-        NfcTag(uint8_t* uid, uint8_t uidLength, String tagType, NdefMessage& ndefMessage);
-        NfcTag(uint8_t* uid, uint8_t uidLength, String tagType, const uint8_t * ndefData, const int ndefDataLength);  
+        NfcTag(byte *uid, unsigned int uidLength);
+        NfcTag(byte *uid, unsigned int uidLength, String tagType);
+        NfcTag(byte *uid, unsigned int uidLength, String tagType, NdefMessage& ndefMessage);
+        NfcTag(byte *uid, unsigned int uidLength, String tagType, const byte *ndefData, const int ndefDataLength);  
         ~NfcTag(void);
         NfcTag& operator=(const NfcTag& rhs);
         uint8_t getUidLength();
-        void getUid(uint8_t* uid, uint8_t uidLength);
+        void getUid(byte *uid, unsigned int uidLength);
         String getUidString();
         String getTagType();
         boolean hasNdefMessage();
         NdefMessage getNdefMessage();
         void print();
     private:
-        uint8_t* _uid;
-        uint8_t _uidLength;
+        byte *_uid;
+        unsigned int _uidLength;
         String _tagType; // Mifare Classic, NFC Forum Type {1,2,3,4}, Unknown
         NdefMessage* _ndefMessage;
         // TODO capacity
